fix(participants folder): move participants folder
