+++ client/src/components/fieldsets/participants/BackgroundInformationFieldset.tsx
@@ -1,15 +1,15 @@
 import { t } from '@lingui/macro'
 import { useLingui } from '@lingui/react'
 import React from 'react'
-import ConditionalCard from '../../../Core/Containers/ConditionalCard'
-import Checkbox from '../../../Core/DataEntry/Checkbox'
-import DateInput from '../../../Core/DataEntry/DateInput'
-import RadioButton from '../../../Core/DataEntry/RadioButton'
-import Select from '../../../Core/DataEntry/Select'
-import Field from '../../../Core/Field/Field'
-import Section from '../../../Core/Field/Section'
-import Column from '../../../Core/Layout/Column/Column'
-import Row from '../../../Core/Layout/Row/Row'
+import ConditionalCard from '../../Core/Containers/ConditionalCard'
+import Checkbox from '../../Core/DataEntry/Checkbox'
+import DateInput from '../../Core/DataEntry/DateInput'
+import RadioButton from '../../Core/DataEntry/RadioButton'
+import Select from '../../Core/DataEntry/Select'
+import Field from '../../Core/Field/Field'
+import Section from '../../Core/Field/Section'
+import Column from '../../Core/Layout/Column/Column'
+import Row from '../../Core/Layout/Row/Row'
 
 interface Props {
     prefillData?: BackgroundInformationFieldsetModel
