+++ client/src/components/fieldsets/participants/MotivationInformationFieldset.tsx
@@ -1,14 +1,14 @@
 import { t } from '@lingui/macro'
 import { useLingui } from '@lingui/react'
 import React from 'react'
-import Checkbox from '../../../Core/DataEntry/Checkbox'
-import Input from '../../../Core/DataEntry/Input'
-import RadioButton from '../../../Core/DataEntry/RadioButton'
-import Field from '../../../Core/Field/Field'
-import Section from '../../../Core/Field/Section'
-import Label from '../../../Core/Label/Label'
-import Column from '../../../Core/Layout/Column/Column'
-import Row from '../../../Core/Layout/Row/Row'
+import Checkbox from '../../Core/DataEntry/Checkbox'
+import Input from '../../Core/DataEntry/Input'
+import RadioButton from '../../Core/DataEntry/RadioButton'
+import Field from '../../Core/Field/Field'
+import Section from '../../Core/Field/Section'
+import Label from '../../Core/Label/Label'
+import Column from '../../Core/Layout/Column/Column'
+import Row from '../../Core/Layout/Row/Row'
 
 interface Props {
     prefillData?: MotivationInformationFieldsetModel
