+++ client/src/components/fieldsets/participants/ReadingTestInformationFieldset.tsx
@@ -1,10 +1,10 @@
 import { t } from '@lingui/macro'
 import { useLingui } from '@lingui/react'
 import React from 'react'
-import Select from '../../../Core/DataEntry/Select'
-import Field from '../../../Core/Field/Field'
-import Section from '../../../Core/Field/Section'
-import Column from '../../../Core/Layout/Column/Column'
+import Select from '../../Core/DataEntry/Select'
+import Field from '../../Core/Field/Field'
+import Section from '../../Core/Field/Section'
+import Column from '../../Core/Layout/Column/Column'
 
 interface Props {
     prefillData?: ReadingTestInformationFieldsetModel
