+++ client/src/views/Authorized/Participants/mocks/participants.ts
@@ -5,14 +5,14 @@ import { ContactInformationFieldsetModel } from '../../../../components/fieldset
 import { CourseInformationFieldsetModel } from '../../../../components/fieldsets/shared/CourseInformationFieldset'
 import { EducationInformationFieldsetModel } from '../../../../components/fieldsets/shared/EducationInformationFieldset'
 import { GeneralInformationFieldsetModel } from '../../../../components/fieldsets/shared/GeneralInformationFieldset'
-import { BackgroundInformationFieldsetModel } from '../../../../components/fieldsets/shared/participants/BackgroundInformationFieldset'
-import { CivicIntegrationFieldsetModel } from '../../../../components/fieldsets/shared/participants/CivicIntegrationInformationFieldset'
-import { LevelInformationFieldsetModel } from '../../../../components/fieldsets/shared/participants/LevelInformationFieldset'
-import { MotivationInformationFieldsetModel } from '../../../../components/fieldsets/shared/participants/MotivationInformationFieldset'
-import { ReadingTestInformationFieldsetModel } from '../../../../components/fieldsets/shared/participants/ReadingTestInformationFieldset'
-import { RefererInformationFieldsetModel } from '../../../../components/fieldsets/shared/participants/ReferrerInformationFieldset'
-import { WorkInformationFieldsetModel } from '../../../../components/fieldsets/shared/participants/WorkInformationFieldset'
-import { WritingInformationFieldsetModel } from '../../../../components/fieldsets/shared/participants/WritingInformationFieldset'
+import { BackgroundInformationFieldsetModel } from '../../../../components/fieldsets/participants/BackgroundInformationFieldset'
+import { CivicIntegrationFieldsetModel } from '../../../../components/fieldsets/participants/CivicIntegrationInformationFieldset'
+import { LevelInformationFieldsetModel } from '../../../../components/fieldsets/participants/LevelInformationFieldset'
+import { MotivationInformationFieldsetModel } from '../../../../components/fieldsets/participants/MotivationInformationFieldset'
+import { ReadingTestInformationFieldsetModel } from '../../../../components/fieldsets/participants/ReadingTestInformationFieldset'
+import { RefererInformationFieldsetModel } from '../../../../components/fieldsets/participants/ReferrerInformationFieldset'
+import { WorkInformationFieldsetModel } from '../../../../components/fieldsets/participants/WorkInformationFieldset'
+import { WritingInformationFieldsetModel } from '../../../../components/fieldsets/participants/WritingInformationFieldset'
 import { PersonInformationFieldsetModel } from '../../../../components/fieldsets/shared/PersonInformationFieldset'
 
 export const taalhuizenParticipantsMock: ParticipantsMock[] = times(3, num => ({
